fix: point HuggingFace home directory to PVC mount
Depending on the distribution configuration, some models might get
downloaded. This is true when the sentence-transformers provider is
used. The library will download embedding models in the local filesystem
using the HuggingFace library. We want to use the PVC mount to store the
Datasets and the Hub artifact. For this, we need to set the `HF_HOME`
env variable to our PVC mount.

This is important so that we:

* Avoid re-downloading models if the Pod moves
* Do not fill the host filesystem
* Take advantage of potentially faster storage than the host

In Kubernetes, when we append environment variables, if there's a
duplicate name, the last one in the list takes precedence. So, a
user-provided environment variable will always be honored.

Signed-off-by: Sébastien Han <seb@redhat.com>