fix: adapt the container entrypoint based on the llama-stack version
In llama-stack 0.2.17, with the help of
https://github.com/llamastack/llama-stack/pull/2975 the module name
containing the server changed. So we need to adapt the entrypoint in
consequence.

When using a ConfigMap to store the server configuration, the operator
overrides the container entrypoint. This code has been adapted to use
the correct module based on the detected llama-stack version.

Also, we now force the entrypoint to use python3 instead of python.

Signed-off-by: Sébastien Han <seb@redhat.com>